+++ imports/components/@enhancers/likes/userLikes.js
@@ -1,0 +1,18 @@
+import { graphql } from "react-apollo";
+import gql from "graphql-tag";
+import { contentCard, groupCard } from "./fragments";
+
+export const LIKES_QUERY = gql`
+  query UserLikes {
+    userFeed(filters:["LIKES"]) {
+      ... ContentCard
+      ... GroupCard
+    }
+  }
+  ${contentCard}
+  ${groupCard}
+`;
+
+export const withLikes = graphql(LIKES_QUERY, { name: "likes", options: { forceFetch: true } });
+
+export default withLikes;