+++ imports/components/people/profile/likes/index.js
@@ -1,55 +1,19 @@
-import { Meteor } from "meteor/meteor";
+
 import { PropTypes } from "react";
-import uniqBy from "lodash.uniqby";
 
-import createContainer from "../../../../deprecated/meteor/react-meteor-data";
-import Likes from "../../../../deprecated/database/collections/likes";
 import Loading from "../../../@primitives/UI/loading";
-import MiniCard from "../../../@primitives/UI/cards/MiniCard";
+import LikesList from "../../../shared/likes-list";
 
-const RenderLikes = ({ likes }) => {
-  if (!likes || !likes.length) return null;
+import withRecentLikes from "../../../@enhancers/likes/recents";
+import withUserLikes from "../../../@enhancers/likes/userLikes";
 
-  return (
-    <div>
-      {
-        likes.map((like, i) => (
-          <div
-            key={i}
-            className="soft-half-bottom@palm-wide"
-            style={{
-              maxWidth: "480px",
-              margin: "0 auto",
-            }}
-          >
-            <MiniCard
-              key={i}
-              title={like.title}
-              content={{
-                channelName: like.category.toLowerCase(),
-                content: {
-                  images: [{
-                    url: like.image,
-                  }],
-                },
-                entryId: like.entryId,
-                icon: like.icon,
-              }}
-              link={like.link || ""}
-            />
-          </div>
-        ))
-      }
-    </div>
-  );
-};
-
-RenderLikes.propTypes = {
-  likes: PropTypes.array,
-};
-
-const RenderRecents = ({ likes, recentLikes }) => {
-  if (likes && likes.length > 0) return null;
+const RenderLikes = ({ likes, recentLikes }) => {
+  if (!likes || !recentLikes || likes.loading || recentLikes.loading) {
+    return <div className="text-center"><Loading /></div>;
+  }
+  if (likes && !likes.loading && likes.userFeed && likes.userFeed.length > 0) {
+    return <LikesList likes={likes.userFeed} />;
+  }
   return (
     <div>
       <p className="soft text-center">
@@ -59,20 +23,17 @@ const RenderRecents = ({ likes, recentLikes }) => {
           </small>
         </em>
       </p>
-      <RenderLikes likes={recentLikes} />
+      <LikesList likes={recentLikes.recentlyLiked} />
     </div>
   );
 };
 
-RenderRecents.propTypes = {
+RenderLikes.propTypes = {
   likes: PropTypes.array,
   recentLikes: PropTypes.array,
 };
 
-// XXX make this dynamic via heighliner
 const LikesContainer = (props) => {
-  if (!props.likes) return <Loading />;
-
   const { likes, recentLikes } = props;
 
   return (
@@ -83,8 +44,7 @@ const LikesContainer = (props) => {
       }
       style={{ marginTop: "-20px" }}
     >
-      <RenderLikes likes={likes} />
-      <RenderRecents likes={likes} recentLikes={recentLikes} />
+      <RenderLikes likes={likes} recentLikes={recentLikes} />
     </div>
   );
 };
@@ -94,26 +54,9 @@ LikesContainer.propTypes = {
   recentLikes: PropTypes.array,
 };
 
-export default createContainer(() => {
-  Meteor.subscribe("likes");
-  const likes = Likes.find({
-    userId: Meteor.userId(),
-  }, { sort: { dateLiked: -1 } }).fetch();
-
-  const recentLikes = Likes.find({
-    userId: {
-      $not: Meteor.userId(),
-    },
-  }, { sort: { dateLiked: -1 } }).fetch();
-
-  return {
-    likes,
-    recentLikes: uniqBy(recentLikes, "entryId"),
-  };
-}, LikesContainer);
-
 export {
   LikesContainer,
-  RenderRecents,
   RenderLikes,
 };
+
+export default withUserLikes(withRecentLikes(LikesContainer));