+++ imports/util/categories/__tests__/categories.name.js
@@ -21,3 +21,13 @@ it("captilizes `articles`", () => {
   const result = categoryName({ channelName: "articles" });
   expect(result).toBe("Articles");
 });
+
+it("returns `Events` for `newspring_now`", () => {
+  const result = categoryName({ channelName: "newspring_now" });
+  expect(result).toBe("Events");
+});
+
+it("returns `Need to Know` for `promotions_newspring`", () => {
+  const result = categoryName({ channelName: "promotions_newspring" });
+  expect(result).toBe("Need to Know");
+});